Fundamentals - Hello, World
Minor refactoring and additional test case showcasing some testing functionalities:
Helper() indicates that in case of errors, point to the where it was called instead
of inside the function.
On composition: the testing.TB is a common interface between testing.T and testing.B
this highlights a bit on leverage has-a relation of interfaces for code reuse.